cmd/flux-ping: handle 32-bit tv_sec/tv_usec
Problem: flux-ping segfaults in flux_rpcf_multi() on arm7l,
which has 32-bit struct timeval members.

The "I" format specifier refers to a 64-bit argument in
json_vpack_ex() and passing 32 bit arguments triggers a segfault.
The segfault is avoided in flux-ping by casting tv_sec and tv_usec
arguments to uint64_t.

+++ src/cmd/flux-ping.c
@@ -176,8 +176,8 @@ void send_ping (struct ping_ctx *ctx)
     rpc = flux_rpcf_multi (ctx->h, ctx->topic, ctx->rank, 0,
                            "{s:i s:I s:I s:s}",
                            "seq", ctx->send_count,
-                           "time.tv_sec", t0.tv_sec,
-                           "time.tv_nsec", t0.tv_nsec,
+                           "time.tv_sec", (uint64_t)t0.tv_sec,
+                           "time.tv_nsec", (uint64_t)t0.tv_nsec,
                            "pad", ctx->pad);
     if (!rpc)
         log_err_exit ("flux_rpcf_multi");
